requirements.txt

psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pandas==2.1.3
numpy==1.24.3
python-dotenv==1.0.0
pyyaml==6.0.1
great-expectations==0.18.3
pytest==7.4.3
black==23.11.0
flake8==6.1.0
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0